Replaced "Sync Changes logic"
Instead of downloading changes and patching the local cache, we now determine if the current directory is up to date and re-download the contents if necessary. To determine if a directory is up to date, we keep a map of id to revision. If a directory is below the current change id, we update it the next time we enter the directory.

Refreshing a folder will recheck the change id and download any changes to the current folder.